+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Automatic academic term selection
 - Tooltip with occupation of the shift (on event hover)
+- Warning of occupation data not in real time
 - CTRL + Click on shift to access course page
 
 ### Fixed